looks like a shit show right now, but i think it's making sense
