+++ debug/queen_test.py
