+++ src/eg.py
@@ -14,6 +14,9 @@ import logging
 import random
 import math
 
+# importing from other folders
+# from folder.folder.file import Whatever
+
 logging.basicConfig(
     level=logging.DEBUG,
     filename="egbot.log",
@@ -93,14 +96,7 @@ class EGbot(sc2.BotAI):
         ):
             larvae.random.train(UnitTypeId.OVERLORD)
 
-    async def opening_strats(self):
-        """
-        Starter method to develop strategies for various openings.  Uses a dictionary with different strategies
-        to select from.  A random number is generated when the program starts determining the strategy.  
-        """
-        # TODO: This is already declared up top and assigned to self.hq. Do we need both?
-        hq: Unit = self.townhalls.first
-        strat_dict = {"pool_first": 1, "expand_first": 2, "double_expand": 3}
+
 
         async def build_pool():  # Build spawning pool
             if (
@@ -131,198 +127,6 @@ class EGbot(sc2.BotAI):
                         if not self.worker_en_route_to_build(UnitTypeId.EXTRACTOR):
                             await self.build(UnitTypeId.EXTRACTOR, vg)
                             break
-
-        # normal strat - let it play out
-        if self.hatch_strat == strat_dict.get("pool_first"):
-            await build_pool()
-            if (
-                self.already_pending(UnitTypeId.SPAWNINGPOOL) == 1
-                and self.structures(UnitTypeId.EXTRACTOR).amount < 1
-            ):
-                await build_gas()
-            await expand()
-            if (
-                self.townhalls.ready.amount == 2
-                and self.structures(UnitTypeId.EXTRACTOR).amount < 3
-            ):
-                await build_gas()
-
-        if self.hatch_strat == strat_dict.get("expand_first"):
-            # expand as long as less than 2 hatcheries
-            if (
-                self.townhalls.ready.amount + self.already_pending(UnitTypeId.HATCHERY)
-                < 2
-            ):
-                await expand()
-            # elif to see if pool AND 2 hatcheries built, if so expand
-            elif (
-                self.structures(UnitTypeId.SPAWNINGPOOL).amount
-                + self.already_pending(UnitTypeId.SPAWNINGPOOL)
-                == 1
-                and self.townhalls.ready.amount
-                + self.already_pending(UnitTypeId.HATCHERY)
-                == 2
-            ):
-                await expand()
-            # if and elif are false, build pool
-            else:
-                await build_pool()
-            # build a gas if pool is pending
-            if (
-                self.already_pending(UnitTypeId.SPAWNINGPOOL) == 1
-                and self.structures(UnitTypeId.EXTRACTOR).amount < 1
-            ):
-                await build_gas()
-            if (
-                self.townhalls.ready.amount == 2
-                and self.structures(UnitTypeId.EXTRACTOR).amount < 3
-            ):
-                await build_gas()
-
-        if self.hatch_strat == strat_dict.get("double_expand"):
-            # expand as long as less than 2 hatcheries
-            if (
-                self.townhalls.ready.amount + self.already_pending(UnitTypeId.HATCHERY)
-                < 3
-            ):
-                await expand()
-            # if pool built AND 3 hatcheries, continue to attempt to expand
-            elif (
-                self.structures(UnitTypeId.SPAWNINGPOOL).amount
-                + self.already_pending(UnitTypeId.SPAWNINGPOOL)
-                == 1
-                and self.townhalls.ready.amount
-                + self.already_pending(UnitTypeId.HATCHERY)
-                == 3
-            ):
-                await expand()
-            # if no pool built yet and 3 hatcheries pending or built, build pool
-            else:
-                await build_pool()
-
-            # build a gas if pool is pending
-            if (
-                self.already_pending(UnitTypeId.SPAWNINGPOOL) == 1
-                and self.structures(UnitTypeId.EXTRACTOR).amount < 1
-            ):
-                await build_gas()
-            if (
-                self.townhalls.ready.amount == 3
-                and self.structures(UnitTypeId.EXTRACTOR).amount < 3
-            ):
-                await build_gas()
-
-    # async def larva_inject(self):
-    #     hatcheries = self.townhalls.ready #list of ready hatcheries
-    #     queens = self.units(UnitTypeId.QUEEN)  # list of queens
-
-    #     # TODO: use queen tags instead
-    #     for hatchery in hatcheries:
-    #         for queen in queens.closer_than(5.0, hatchery):
-    #             if queen.energy >= 25:
-    #                 queen(AbilityId.EFFECT_INJECTLARVA, hatchery)
-
-    async def spread_creep(self):
-
-        positions = []
-        filtered_locations = []
-        build_tumor = AbilityId.BUILD_CREEPTUMOR_QUEEN
-        unused_tumors = []
-        enemy_base = self.enemy_start_locations
-
-        if hasattr(self, "queens_assigned_hatcheries"):
-            unassignedQueens = self.units(UnitTypeId.QUEEN).filter(
-                lambda q: (
-                    q.tag not in self.queens_assigned_hatcheries
-                    and q.energy >= 25
-                    or q.energy >= 50
-                )
-                and (
-                    q.is_idle
-                    or len(q.orders) == 1
-                    and q.orders[0].ability.id in [AbilityId.MOVE]
-                )
-            )
-        else:
-            unassignedQueens = self.units(UnitTypeId.QUEEN).filter(
-                lambda q: q.energy >= 25
-                and (
-                    q.is_idle
-                    or len(q.orders) == 1
-                    and q.orders[0].ability.id in [AbilityId.MOVE]
-                )
-            )
-
-        # queen
-        for queen in unassignedQueens:
-            positions = self.get_pos_around_unit(
-                queen, min_range=5, max_range=10, loc_amt=12
-            )
-
-            # filter out places without creep
-            for loc in positions:
-                if self.has_creep(loc):
-                    filtered_locations.append(loc)
-
-                # find shortest distance to enemy base
-            shortest_distance = 1000.00
-            for loc in filtered_locations:
-                temp_distance = math.sqrt(
-                    (enemy_base[0].x - loc.x) ** 2.0 + (enemy_base[0].y - loc.y) ** 2.0
-                )
-                if temp_distance < shortest_distance:
-                    best_loc = loc
-                    shortest_distance = temp_distance
-
-            queen(build_tumor, best_loc)
-
-        """TODO: Determine directions for creep spread - still random, but only spreads outwards away from starting hatchery
-            TODO: Ensure creep doesn't spread onto hatchery locations - potentially done"""
-
-        filtered_locations = []  # locations to place tumors
-        # all possible tumors
-        tumors = (
-            self.structures(UnitTypeId.CREEPTUMORQUEEN)
-            | self.structures(UnitTypeId.CREEPTUMOR)
-            | self.structures(UnitTypeId.CREEPTUMORBURROWED)
-        )
-
-        if tumors:  # if tumors exists
-            all_tumors_abilities = await self.get_available_abilities(
-                tumors
-            )  # get all the abilities from available tumors
-            for i, abilities in enumerate(
-                all_tumors_abilities
-            ):  # loop through tumors that have abilities
-                tumor = tumors[i]  # select single tumor
-                if not tumor.is_idle and isinstance(
-                    tumor.order_target, Point2
-                ):  # if tumor is in process of morphing or something, add to used.tumors
-                    self.used_tumors.add(tumor.tag)
-                    continue
-                # ensure tumor has Ability to spread creep
-                if AbilityId.BUILD_CREEPTUMOR_TUMOR in abilities:
-                    # gets positions around tumor - note: loc_amt does not seem to work, returned over 70 positions last I checked
-                    # max range seems too far as well - was getting out of range errors
-                    positions = self.get_pos_around_unit(
-                        tumor, min_range=5, max_range=20, loc_amt=12
-                    )
-                    # determine which positions have creep
-                    for loc in positions:
-                        if self.has_creep(loc) and not self._position_blocks_expansion(
-                            loc
-                        ):
-                            # working towards sending creep out - this filters positions that are closer to the starting hatchery, so tumors don't go backwards, can go sideways though
-                            if self._distance_pos_to_pos(
-                                tumor.position, self.townhalls.first.position
-                            ) <= self._distance_pos_to_pos(
-                                loc, self.townhalls.first.position
-                            ):
-                                filtered_locations.append(loc)
-
-                    pos = random.randrange(0, len(filtered_locations))
-                    # build that tumor!
-                    tumor(AbilityId.BUILD_CREEPTUMOR_TUMOR, filtered_locations[pos])
 
     def get_pos_around_unit(
         self, unit, min_range=0, max_range=500, step_size=1, loc_amt=32
@@ -364,7 +168,6 @@ class EGbot(sc2.BotAI):
                 blocks_expansion = True
                 break
         return blocks_expansion
-
 
     async def on_building_construction_complete(self, unit: Unit):
         # TODO: Possibly where we can create Queens upon building completion.
