+++ src/managers/combat/micro.py
@@ -1,0 +1,1 @@
+"""A manager that would deal with microing units in combat"""
